tests/devops_tests/requirements.txt
pytest
ghapi
gitpython
binaryornot
